(maint) Set `gpg_key` for the release user
With our attempt to ensure we're signing all packages with the same key,
we should set the gpg_key here. This ensures we use the same key for all
projects. If a projects forgets to set `gpg_key`, then we can still
successfully sign packages. If a projects wants to use a different key,
then they can set it in the project-specific build-data file.