(maint) Add ips_signing_ssh_key
This commit adds and sets the `ips_signing_ssh_key` param. This is used when
sshing into our solaris signer. Currently, this param is set as an environment
variable in several jenkins jobs. Adding it here will improve maintainability,
since we won't have to add this param to every job that does solaris signing.