(RE-13743) 2nd try at creating correct 'freight cache' for apt repos
`freight cache` is very particular about how things are structured. If
we want to have separate apt repos for each puppet major
version (puppet7, puppet8, etc.) we need to only allow freight to deal
with one at a time.

The structure of the freight VARLIB needs to be like
`/opt/tools/freight/<puppet-version>` which allows us to call `freight
cache apt/<distro>` and expect it to come out properly.